requirements.txt
# Web Framework
flask>=2.0.0
flask-cors>=4.0.0
gunicorn>=20.1.0
Werkzeug>=2.0.0
openai

# Configuration
python-dotenv>=0.19.0

# Development Tools
black==23.3.0         # format
ruff==0.0.270        # check
pytest>=7.0.0        # test
locust

# Optional - if need data processing
numpy>=1.21.0
pandas>=1.3.0
psycopg2-binary>=2.9.0